+++ apidaora/openapi/path_decorator/factories/path_args.py
@@ -4,8 +4,7 @@ from typing import Any, Dict, Type, _TypedDictMeta  # type: ignore
 from jsondaora import jsondaora
 
 from ....core.request import PathArgs
-from ....core.router import split_path
-from ....core.regex_router import PATH_RE
+from ....core.router import PATH_RE, split_path
 from ....exceptions import InvalidPathParams
